preliminary fault detection code
